+++ objects.h
@@ -29,7 +29,6 @@ extern XColor color_resize_window;
 extern XColor color_minimize_window;
 extern XColor color_maximize_window;
 extern XColor color_close_window;
-extern XColor color_exit;
 
 extern GC gc_text_white;
 extern GC gc_text_black;