+++ utils.h
@@ -11,7 +11,8 @@ typedef enum utils_t {
     UTILS_LOGO,
     UTILS_WALLPAPER_PNG,
     UTILS_WALLPAPER_JPG,
-	UTILS_WALLPAPER_TO_USE
+	UTILS_WALLPAPER_TO_USE,
+	UTILS_LOGOUT
 } utils_t;
 
 char* utils_get(utils_t util);